+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.2.0] - 2021-03-31
+
+### Changed
+
+- Renamed Health Monitoring to Resource Monitoring.
+
+### Removed
+
+- Moved Resource Monitoring to the KF-Util library.
+ 
 ## [2.1.1] - 2025-03-27
 
 ### Fixed